+++ documentation/src/main/wiki/WhatsNew.md
@@ -1,3 +1,3 @@
 # What's new
 
-* View Knot separated to Service Knot and Handlebars Knot
+* Simplified JSON configuration. Knot.x Application should have minimal config with list of services with ability to override default settings in this config and through system properties.